Move/Rename: EncodingManager.Access -> WayAccess
