+++ core/src/test/java/com/graphhopper/routing/util/RacingBikeTagParserTest.java
@@ -28,8 +28,8 @@ import com.graphhopper.util.PMap;
 import org.junit.jupiter.api.Test;
 
 import static com.graphhopper.routing.util.BikeCommonTagParser.PUSHING_SECTION_SPEED;
-import static com.graphhopper.routing.util.EncodingManager.Access.WAY;
 import static com.graphhopper.routing.util.PriorityCode.*;
+import static com.graphhopper.routing.util.WayAccess.WAY;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -223,7 +223,7 @@ public class RacingBikeTagParserTest extends AbstractBikeTagParserTester {
         EncodingManager encodingManager = EncodingManager.create("racingbike");
         BikeCommonTagParser parser = new RacingBikeTagParser(encodingManager, new PMap("block_fords=true")) {
             @Override
-            public EncodingManager.Access getAccess(ReaderWay way) {
+            public WayAccess getAccess(ReaderWay way) {
                 return WAY;
             }
         };
